# [Performance] RPS Improvement +500 RPS when sending the `user` field