## Cache exception handled 


## Changes

### Short Explanation
- Standardized the caching logic to avoid type issues, which was causing 50% cache misses.

### Context
The set_cache function was storing values with unexpected types. As a result, the get_cache function failed to recognize objects that were actually present in the cache about half the time. This caused unnecessary database lookups roughly 50% of the time, leading to the performance regression shown below.

### Performance Change

<p><strong>Before Fix</strong></p>

Type | Name | # Requests | # Fails | Median (ms) | 95%ile (ms) | 99%ile (ms) | Average (ms) | Min (ms) | Max (ms) | Average size (bytes) | Current RPS | Current Failures/s
-- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | --
POST | /chat/completions | 31423 | 0 | 27 | 6200 | 11000 | 1212.85 | 12 | 26398 | 398 | 558.7 | 0
  | Aggregated | 31423 | 0 | 27 | 6200 | 11000 | 1212.85 | 12 | 26398 | 398 | 558.7 | 0

<p>After the Fix</p>

Type | Name | # Requests | # Fails | Median (ms) | 95%ile (ms) | 99%ile (ms) | Average (ms) | Min (ms) | Max (ms) | Average size (bytes) | Current RPS | Current Failures/s
-- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | --
POST | /chat/completions | 36496 | 0 | 86 | 840 | 13000 | 488.78 | 15 | 15683 | 398 | 1039.4 | 0
Custom | LiteLLM Overhead Duration (ms) | 36496 | 0 | 4 | 52 | 190 | 14.13 | 0 | 626 | 0 | 1039.4 | 0
  | Aggregated | 72992 | 0 | 31 | 480 | 10000 | 251.46 | 0 | 15683 | 199 | 2078.8 | 0


<!-- notionvc: f891e75d-c376-4a0a-a63c-60f3a3ed8506 -->


## Relevant issues

None

## Pre-Submission checklist

**Please complete all items before asking a LiteLLM maintainer to review your PR**

- [ ] I have Added testing in the [`tests/litellm/`](https://github.com/BerriAI/litellm/tree/main/tests/litellm) directory, **Adding at least 1 test is a hard requirement** - [see details](https://docs.litellm.ai/docs/extras/contributing_code)
- [ ] I have added a screenshot of my new test passing locally 
- [x] My PR passes all unit tests on [`make test-unit`](https://docs.litellm.ai/docs/extras/contributing_code)
- [x] My PR's scope is as isolated as possible, it only solves 1 specific problem


## Type

🐛 Bug Fix
🧹 Refactoring